build-632-review-comments :
1. Hint in Hebrew is not in the right place. Please for both Hebrew &
English bring the hint inside the frame (and not to make the frame
bigger), as per the attached image (now vs. how it should be for both
English & Hebrew).

2. Practice Questions - the lesson should start directly from first
question (i.e. skip lesson intro, same as with "start over"). 

4. Make font size of "hint' in "play lesson" bigger than it is today
(it's now very small). I think as big as it is in "preview question"
would be good.